backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
PyMySQL==1.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<5
twilio>=8.0.0

# Face Recognition
face_recognition>=1.3.0
opencv-python>=4.8
numpy>=1.24,<2
Pillow>=10.0
# High-accuracy models
insightface>=0.7.0
deepface>=0.0.79
retina-face>=0.0.17
onnxruntime>=1.16.0
tensorflow>=2.15
tf-keras>=2.20  # Required by retina-face with TensorFlow 2.20+

# File Handling
python-multipart==0.0.6
aiofiles==23.2.1

# Validation & Config
pydantic==2.5.3
email-validator>=2.0
pydantic-settings==2.1.0

# CORS
starlette==0.35.1
